+++ include/trifinger_object_tracking/pybullet_tricamera_object_tracker_driver.hpp
@@ -11,6 +11,9 @@
 
 #include <robot_interfaces/finger_types.hpp>
 #include <robot_interfaces/sensors/sensor_driver.hpp>
+#include <trifinger_cameras/camera_parameters.hpp>
+#include <trifinger_cameras/pybullet_tricamera_driver.hpp>
+
 #include <trifinger_object_tracking/tricamera_object_observation.hpp>
 
 namespace trifinger_object_tracking
@@ -22,7 +25,8 @@ namespace trifinger_object_tracking
  */
 class PyBulletTriCameraObjectTrackerDriver
     : public robot_interfaces::SensorDriver<
-          trifinger_object_tracking::TriCameraObjectObservation>
+          trifinger_object_tracking::TriCameraObjectObservation,
+          trifinger_cameras::TriCameraInfo>
 {
 public:
     // For some reason the constructor needs to be in the header file to avoid a
@@ -41,52 +45,30 @@ public:
     PyBulletTriCameraObjectTrackerDriver(
         pybind11::object tracking_object,
         robot_interfaces::TriFingerTypes::BaseDataPtr robot_data,
-        bool render_images = true)
-        : tracking_object_(tracking_object),
-          robot_data_(robot_data),
-          render_images_(render_images),
-          last_update_robot_time_index_(0)
-    {
-        // initialize Python interpreter if not already done
-        if (!Py_IsInitialized())
-        {
-            pybind11::initialize_interpreter();
-        }
+        bool render_images = true,
+        trifinger_cameras::Settings settings = trifinger_cameras::Settings());
 
-        pybind11::gil_scoped_acquire acquire;
-
-        if (render_images)
-        {
-            // some imports that are needed later for converting images (numpy
-            // array -> cv::Mat)
-            numpy_ = pybind11::module::import("numpy");
-
-            // TriFingerCameras gives access to the cameras in simulation
-            pybind11::module mod_camera =
-                pybind11::module::import("trifinger_simulation.camera");
-            cameras_ = mod_camera.attr("TriFingerCameras")();
-        }
-    }
+    /**
+     * @brief Get the camera parameters.
+     *
+     * @verbatim embed:rst:leading-asterisk
+     * Internally, this calls
+     * :cpp:func:`trifinger_cameras::PyBulletTriCameraDriver::get_sensor_info`,
+     * so see there fore more information.
+     * @endverbatim
+     */
+    trifinger_cameras::TriCameraInfo get_sensor_info() override;
 
     //! @brief Get the latest observation.
-    trifinger_object_tracking::TriCameraObjectObservation get_observation();
+    trifinger_object_tracking::TriCameraObjectObservation get_observation()
+        override;
 
 private:
-    //! @brief Python object to access cameras in pyBullet.
-    pybind11::object cameras_;
+    //! @brief PyBullet driver instance for rendering images.
+    trifinger_cameras::PyBulletTriCameraDriver camera_driver_;
 
     //! @brief Python object of the tracked object.
     pybind11::object tracking_object_;
-
-    robot_interfaces::TriFingerTypes::BaseDataPtr robot_data_;
-
-    //! @brief If false, no actual images are rendered.
-    bool render_images_;
-
-    //! @brief The numpy module.
-    pybind11::module numpy_;
-
-    time_series::Index last_update_robot_time_index_;
 };
 
 }  // namespace trifinger_object_tracking